Remove the template from the data buffer
