+++ CMakeLists.txt
@@ -9,13 +9,14 @@ add_executable(${PROJECT_NAME})
 add_compile_options(-Wall -Wextra -Werror -O2)
 
 find_package(Curses REQUIRED)
+
 target_include_directories(${PROJECT_NAME} PRIVATE ${CURSES_INCLUDE_DIR})
 target_include_directories(${PROJECT_NAME} PRIVATE .)
 
 target_sources(${PROJECT_NAME} PRIVATE
 	hexit.cc
 	terminal_window.cc
+	data_buffer.cc
 )
-
 
 target_link_libraries(${PROJECT_NAME} ${CURSES_LIBRARIES})
